pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icon-gen-ai"
version = "0.1.5"
description = "Generate customizable icons from Iconify with easy export to PNG, SVG, WebP"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Yauheniya Varabyova", email = "yauheniya.ai@gmail.com" }
]
keywords = ["icons", "logos", "iconify", "svg", "png", "webp"]

dependencies = [
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "cairosvg>=2.7.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
icon-gen-ai = "icon_gen.cli:cli"

[project.urls]
Homepage = "https://github.com/yauheniya-ai/icon-gen"
Documentation = "https://github.com/yauheniya-ai/icon-gen#readme"
Repository = "https://github.com/yauheniya-ai/icon-gen"
"Bug Tracker" = "https://github.com/yauheniya-ai/icon-gen/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "black",
    "pytest",
    "pytest-cov",
    "ruff",
]
